pyproject.toml
[project]
name = "secure-finetune"
version = "1.0.0"
description = "Iterative SLM fine-tuning framework with LLM judge feedback"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.36",
    "datasets>=2.14",
    "peft>=0.7",
    "trl>=0.7",
    "bitsandbytes>=0.41",
    "accelerate>=0.25",
    "scikit-learn>=1.3",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "anthropic>=0.18",
    "sentencepiece>=0.1.99",
    "protobuf>=3.20",
]

[project.scripts]
secure-finetune = "secure_finetune.main:main"
